requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
aiokafka==0.11.0
psycopg2-binary==2.9.9
prometheus-client==0.20.0
python-multipart==0.0.9
httpx==0.27.2
scikit-learn==1.5.2
numpy==2.0.2
ruff==0.6.9
pytest==8.3.3
